+++ sdk/nusantara/include/visitor/context/operasi/operasi_sama_dengan_context.h
@@ -9,17 +9,18 @@
 class OperasiSamaDenganContext: public Context {
   public:
     explicit OperasiSamaDenganContext(
-        nstd::daftar<std::unique_ptr<Context>> kumpulanOperasiAtauContext,
+        nstd::daftar<std::unique_ptr<Context>>
+            kumpulanOperasiKondisionalContext,
         nstd::daftar<Token> kumpulanOperator
     );
     static OperasiSamaDenganContext generate(
         const nstd::daftar<std::unique_ptr<ParserTree>>& children
     );
     [[nodiscard]] const nstd::daftar<std::unique_ptr<Context>>&
-    getKumpulanOperasiAtauContext() const;
+    getKumpulanOperasiKondisionalContext() const;
     [[nodiscard]] const nstd::daftar<Token>& getKumpulanOperator() const;
 
   private:
-    nstd::daftar<std::unique_ptr<Context>> kumpulanOperasiAtauContext;
+    nstd::daftar<std::unique_ptr<Context>> kumpulanOperasiKondisionalContext;
     nstd::daftar<Token> kumpulanOperator;
 };