+++ sdk/nusantara/src/parser/parser.cc
@@ -18,20 +18,12 @@ Parser::Parser(ErrorInfo errorInfo, Lexer lexer):
     currentToken(this->lexer.getNextToken()) {}
 
 void Parser::skipWs() {
-	while(
-			matchOr(
-				{
-					TokenType::SPASI, 
-					TokenType::TAB, 
-					TokenType::BARIS_BARU,
-          TokenType::OTHER_WHITESPACE, 
-					TokenType::KOMENTAR_BANYAK_BARIS,
-          TokenType::KOMENTAR_SATU_BARIS, 
-					TokenType::KOMENTAR_DOKUMENTASI
-				}
-  		)
-	) {
-		this->currentToken = this->lexer.getNextToken();
+  while(matchOr(
+      {TokenType::SPASI, TokenType::TAB, TokenType::BARIS_BARU,
+       TokenType::OTHER_WHITESPACE, TokenType::KOMENTAR_BANYAK_BARIS,
+       TokenType::KOMENTAR_SATU_BARIS, TokenType::KOMENTAR_DOKUMENTASI}
+  )) {
+    this->currentToken = this->lexer.getNextToken();
   }
 }
 
@@ -39,10 +31,10 @@ std::unique_ptr<ParserTree> Parser::eat(
     const TokenType& type, const bool& skipWs
 ) {
   if(this->currentToken.getType() == type) {
-		const Token beforeToken = this->currentToken;
+    const Token beforeToken = this->currentToken;
     this->currentToken = this->lexer.getNextToken();
-		if(skipWs) {this->skipWs();}
-		if(match(TokenType::TIDAK_DIKENALI)) {
+    if(skipWs) { this->skipWs(); }
+    if(match(TokenType::TIDAK_DIKENALI)) {
       throw std::runtime_error(this->errorInfo.inLine(
           this->currentToken, "Karakter tidak di kenali."
       ));
@@ -121,7 +113,7 @@ std::unique_ptr<ParserTree> Parser::parseNusantara() {
     nusantara->addChild(this->parseEkspresi());
     nusantara->addChild(eat(TokenType::TITIK_KOMA));
   }
-	nusantara->addChild(eat(TokenType::AKHIR_DARI_FILE));
+  nusantara->addChild(eat(TokenType::AKHIR_DARI_FILE));
   return nusantara;
 }
 
@@ -215,8 +207,22 @@ std::unique_ptr<ParserTree> Parser::parseOperasiKaliSamaDengan() {
 std::unique_ptr<ParserTree> Parser::parseOperasiSamaDengan() {
   return this->fragmentMultiOperasiLeftRight(
       ParserRule::operasi_sama_dengan,
-      [this]() { return this->parseOperasiAtau(); }, TokenType::SAMA_DENGAN
+      [this]() { return this->parseOperasiKondisional(); },
+      TokenType::SAMA_DENGAN
   );
+}
+
+std::unique_ptr<ParserTree> Parser::parseOperasiKondisional() {
+  std::unique_ptr<ParserTree> operasiKondisional =
+      std::make_unique<ParserRuleTree>(ParserRule::operasi_kondisional);
+  operasiKondisional->addChild(this->parseOperasiAtau());
+  if(match(TokenType::TANDA_TANYA)) {
+    operasiKondisional->addChild(this->eat(TokenType::TANDA_TANYA));
+    operasiKondisional->addChild(this->parseEkspresi());
+    operasiKondisional->addChild(this->eat(TokenType::TITIK_DUA));
+    operasiKondisional->addChild(this->parseEkspresi());
+  }
+  return operasiKondisional;
 }
 
 std::unique_ptr<ParserTree> Parser::parseOperasiAtau() {
@@ -388,7 +394,7 @@ std::unique_ptr<ParserTree> Parser::parseNilai() {
   const nstd::daftar<TokenType> types = {
       TokenType::BILANGAN, TokenType::BENAR, TokenType::SALAH
   };
-	this->skipWs();
+  this->skipWs();
   if(this->matchOr(types)) {
     for(const TokenType& type : types) {
       if(this->match(type)) {