+++ sdk/nusantara/src/visitor/context/operasi/operasi_sama_dengan_context.cc
@@ -1,25 +1,27 @@
 #include "visitor/context/operasi/operasi_sama_dengan_context.h"
 
-#include "visitor/context/operasi/operasi_atau_context.h"
+#include "visitor/context/operasi/operasi_kondisional_context.h"
 
 OperasiSamaDenganContext::OperasiSamaDenganContext(
-    nstd::daftar<std::unique_ptr<Context>> kumpulanOperasiAtauContext,
+    nstd::daftar<std::unique_ptr<Context>> kumpulanOperasiKondisionalContext,
     nstd::daftar<Token> kumpulanOperator
 ):
-    kumpulanOperasiAtauContext(std::move(kumpulanOperasiAtauContext)),
+    kumpulanOperasiKondisionalContext(
+        std::move(kumpulanOperasiKondisionalContext)
+    ),
     kumpulanOperator(std::move(kumpulanOperator)) {}
 
 OperasiSamaDenganContext OperasiSamaDenganContext::generate(
     const nstd::daftar<std::unique_ptr<ParserTree>>& children
 ) {
-  nstd::daftar<std::unique_ptr<Context>> kumpulanOperasiAtauContext;
+  nstd::daftar<std::unique_ptr<Context>> kumpulanOperasiKondisionalContext;
   nstd::daftar<Token> kumpulanOperator;
   for(const std::unique_ptr<ParserTree>& child : children) {
     if(auto* prt = dynamic_cast<ParserRuleTree*>(child.get())) {
-      if(prt->getRule() == ParserRule::operasi_atau) {
-        kumpulanOperasiAtauContext.push_back(
-            std::make_unique<OperasiAtauContext>(
-                OperasiAtauContext::generate(prt->getChildren())
+      if(prt->getRule() == ParserRule::operasi_kondisional) {
+        kumpulanOperasiKondisionalContext.push_back(
+            std::make_unique<OperasiKondisionalContext>(
+                OperasiKondisionalContext::generate(prt->getChildren())
             )
         );
       }
@@ -30,13 +32,13 @@ OperasiSamaDenganContext OperasiSamaDenganContext::generate(
     }
   }
   return OperasiSamaDenganContext(
-      std::move(kumpulanOperasiAtauContext), std::move(kumpulanOperator)
+      std::move(kumpulanOperasiKondisionalContext), std::move(kumpulanOperator)
   );
 }
 
 const nstd::daftar<std::unique_ptr<Context>>&
-OperasiSamaDenganContext::getKumpulanOperasiAtauContext() const {
-  return this->kumpulanOperasiAtauContext;
+OperasiSamaDenganContext::getKumpulanOperasiKondisionalContext() const {
+  return this->kumpulanOperasiKondisionalContext;
 }
 
 const nstd::daftar<Token>& OperasiSamaDenganContext::getKumpulanOperator(